allow deleting resource files and adding tags

+++ partners/digital_campus_publish.py
@@ -28,47 +28,48 @@ MPOWERING_DEFAULT_TAGS = ["Digital Campus",
                           "Tablet", 
                           "Creative Commons 2.0 (CC-BY-NC-SA)"]
 
-from api import orb_api, orb_resource, orb_resource_file, orb_resource_url
+from api import orb_api, orb_resource, orb_resource_file, orb_resource_url, ORBAPIResourceExistsException
 
 
-def run(orb_url, orb_username, orb_key, db_name, db_user, db_passwd): 
+def run(orb_url, orb_username, orb_key, db_name, db_user, db_passwd, update_files): 
     api = orb_api()
     api.base_url = orb_url
     api.user_name = orb_username
     api.api_key = orb_key
     
-    
     db = MySQLdb.connect(host="localhost",
-                     user=db_user, 
-                      passwd=db_passwd,
-                      db=db_name)
-    
+                         user=db_user, 
+                          passwd=db_passwd,
+                          db=db_name)
+        
     db.autocommit(True)
 
     # you must create a Cursor object. It will let
     #  you execute all the queries you need
     cur = db.cursor(MySQLdb.cursors.DictCursor) 
-    '''
-    Update with the most recent Moodle backup versions
-    '''
-    cur.execute("SELECT f.id, course_id, f.description, file, moodle_course_id, type, short_name, c.location_code FROM dc_file f INNER JOIN dc_course c ON c.id = f.course_id")
-    for row in cur.fetchall():
-        course_backups = glob.glob(MOODLE_BACKUP_DIR+'backup-moodle2-course-'+str(row['moodle_course_id'])+'-*.mbz')
-        course_backups.sort()
-        file_name = course_backups[len(course_backups)-1]
-        file_date = re.findall(r'[0-9]{8}', file_name)
-        
-        old_course_backups = glob.glob(os.path.join(OUTPUT_DIR_BASE, row['location_code'], 'moodle-backups', row['short_name'] + '-' + row['type'] + '*.mbz'))
-        for old_file in old_course_backups:
-            print "Removing: " + old_file
-            os.remove(old_file)
     
-        new_file_name = row['short_name'] + "-" + row['type'] + "-" + file_date[0] +".mbz"
-        print "Copying over: " + new_file_name
-        shutil.copy2(file_name, os.path.join(OUTPUT_DIR_BASE, row['location_code'], 'moodle-backups', new_file_name))
+    if update_files == 'True':
+        '''
+        Update with the most recent Moodle backup versions
+        '''
+        cur.execute("SELECT f.id, course_id, f.description, file, moodle_course_id, type, short_name, c.location_code FROM dc_file f INNER JOIN dc_course c ON c.id = f.course_id")
+        for row in cur.fetchall():
+            course_backups = glob.glob(MOODLE_BACKUP_DIR+'backup-moodle2-course-'+str(row['moodle_course_id'])+'-*.mbz')
+            course_backups.sort()
+            file_name = course_backups[len(course_backups)-1]
+            file_date = re.findall(r'[0-9]{8}', file_name)
+            
+            old_course_backups = glob.glob(os.path.join(OUTPUT_DIR_BASE, row['location_code'], 'moodle-backups', row['short_name'] + '-' + row['type'] + '*.mbz'))
+            for old_file in old_course_backups:
+                print "Removing: " + old_file
+                os.remove(old_file)
         
-        #cur2 = db.cursor(MySQLdb.cursors.DictCursor)
-        cur.execute("""UPDATE dc_file SET file = '%s' WHERE id = %s """ % (new_file_name, int(row['id'])))
+            new_file_name = row['short_name'] + "-" + row['type'] + "-" + file_date[0] +".mbz"
+            print "Copying over: " + new_file_name
+            shutil.copy2(file_name, os.path.join(OUTPUT_DIR_BASE, row['location_code'], 'moodle-backups', new_file_name))
+            
+            #cur2 = db.cursor(MySQLdb.cursors.DictCursor)
+            cur.execute("""UPDATE dc_file SET file = '%s' WHERE id = %s """ % (new_file_name, int(row['id'])))
         
     '''
     Publish updates to the mPowering
@@ -87,57 +88,70 @@ def run(orb_url, orb_username, orb_key, db_name, db_user, db_passwd):
             resource.study_time_number = row['study_hours']
             resource.study_time_unit = 'hours' 
         
-        resource.id = api.add_resource(resource)
+        update = True
+        try:
+            resource.id = api.add_resource(resource)
+            update = False
+        except ORBAPIResourceExistsException, e:
+            print e.message + ", id no:" + str(e.pk)
+            resource.id = e.pk  
+             
         '''
         If resource has been created
         '''
-        if resource.id:
+        if not update:
             # add resource icon
             if row['icon']:
-                api.add_resource_image(resource.id,os.path.join(IMAGE_DIR_BASE, row['icon']))
-                
-            # add all the default tags
-            for tag in MPOWERING_DEFAULT_TAGS:
-                api.add_resource_tag(resource.id, tag.strip())
-            
-            # add resource specific tags
-            if row['tags']:
-                tag_list = [x.strip() for x in row['tags'].split(',')]
-                for tag in tag_list:
-                    api.add_resource_tag(resource.id, tag.strip())
-             
-            # add the files
-            cur.execute("""SELECT * FROM dc_file WHERE course_id = %s """ % (row['id']))
-            for file in cur.fetchall():
-                resource_file = orb_resource_file()
-                resource_file.title = file['title']
-                resource_file.description = file['description']
-                resource_file.order_by = file['orderby']
-                resource_file.file = os.path.join(OUTPUT_DIR_BASE, row['location_code'], 'moodle-backups', file['file']) 
-                api.add_resource_file(resource.id,resource_file)
-                
-            # add the urls
-            cur.execute("""SELECT * FROM dc_url WHERE course_id = %s """ % (row['id']))
-            for url in cur.fetchall():
-                resource_url = orb_resource_url()
-                resource_url.title = url['title']
-                resource_url.description = url['description']
-                resource_url.order_by = url['orderby']
-                resource_url.url = url['url']
-                api.add_resource_url(resource.id,resource_url)
+                api.add_resource_image(resource.id,os.path.join(IMAGE_DIR_BASE, row['icon']))  
         else: 
             '''
             if resource should be updated
             '''
             # get the resource id
-            resource_api = api.get_resource(resource)
+            resource_from_api = api.get_resource(resource)
             
             # update the resource
             
+            # remove all ResourceFiles
+            for f in resource_from_api['files']:
+                api.delete_resource_file(f['resource_uri'])
+                
+            
+            # remove all ResourceURLs
+            
+            
             # remove all tags for resource
             
-            # add new tags
+         
+        # add all the default tags
+        for tag in MPOWERING_DEFAULT_TAGS:
+            api.add_resource_tag(resource.id, tag.strip())
+        
+        # add resource specific tags
+        if row['tags']:
+            tag_list = [x.strip() for x in row['tags'].split(',')]
+            for tag in tag_list:
+                api.add_resource_tag(resource.id, tag.strip())
+         
+        # add the files
+        cur.execute("""SELECT * FROM dc_file WHERE course_id = %s """ % (row['id']))
+        for file in cur.fetchall():
+            resource_file = orb_resource_file()
+            resource_file.title = file['title']
+            resource_file.description = file['description']
+            resource_file.order_by = file['orderby']
+            resource_file.file = os.path.join(OUTPUT_DIR_BASE, row['location_code'], 'moodle-backups', file['file']) 
+            api.add_resource_file(resource.id,resource_file)
             
+        # add the urls
+        cur.execute("""SELECT * FROM dc_url WHERE course_id = %s """ % (row['id']))
+        for url in cur.fetchall():
+            resource_url = orb_resource_url()
+            resource_url.title = url['title']
+            resource_url.description = url['description']
+            resource_url.order_by = url['orderby']
+            resource_url.url = url['url']
+            api.add_resource_url(resource.id,resource_url)   
             
     cur.close()
     
@@ -149,5 +163,6 @@ if __name__ == "__main__":
     parser.add_argument("db_name", help="Database name")
     parser.add_argument("db_user", help="Database user name")
     parser.add_argument("db_passwd", help="Database password")
+    parser.add_argument("update_files", help="Update files True/False")
     args = parser.parse_args()
-    run(args.orb_url, args.orb_username, args.orb_key, args.db_name, args.db_user, args.db_passwd) 
+    run(args.orb_url, args.orb_username, args.orb_key, args.db_name, args.db_user, args.db_passwd, args.update_files)
